refactor: cast block height and count to long in P2PDeviceApi
The code in 'P2PDeviceApi.cs' has been updated to cast 'block.Height' and 'blockCount' to long before comparing them. This change ensures correct results when dealing with larger number values, thereby preventing potential synchronisation issues within the peer-to-peer network.

+++ core/Network/P2PDeviceApi.cs
@@ -99,7 +99,7 @@ public class P2PDeviceApi : IP2PDeviceApi
             var block = MessagePackSerializer.Deserialize<Models.Block>(parameters[0].Value);
             var blockCount = _systemCore.UnitOfWork().HashChainRepository.Count;
 
-            if (block.Height - blockCount >= 3)
+            if ((long)block.Height - (long)blockCount >= 3)
             {
                 await _systemCore.Sync().SynchronizeAsync();
             }
